+++ src/app/(admin)/admin/dashboard/capsules/page.tsx
@@ -1,5 +1,5 @@
 import CapsuleSection from "@/components/dashboard/admin/contents/page/CapsuleSection";
-import AdminDashboardPageSkeleton from "@/components/skeleton/admin/AdminDashboardPageSkeleton";
+import AdminDashboardPageSkeleton from "@/components/ui/skeleton/admin/AdminDashboardPageSkeleton";
 import { Suspense } from "react";
 
 export default function Page() {
