+++ src/app/(admin)/admin/dashboard/moderation-logs/page.tsx
@@ -1,5 +1,5 @@
 import ModerationSection from "@/components/dashboard/admin/contents/page/ModerationSection";
-import AdminDashboardPageSkeleton from "@/components/skeleton/admin/AdminDashboardPageSkeleton";
+import AdminDashboardPageSkeleton from "@/components/ui/skeleton/admin/AdminDashboardPageSkeleton";
 import { Suspense } from "react";
 
 export default function Page() {
