+++ src/app/(admin)/admin/dashboard/phone-verifications/page.tsx
@@ -1,5 +1,5 @@
 import PhoneVerificationSection from "@/components/dashboard/admin/contents/page/PhoneVerificationSection";
-import AdminDashboardPageSkeleton from "@/components/skeleton/admin/AdminDashboardPageSkeleton";
+import AdminDashboardPageSkeleton from "@/components/ui/skeleton/admin/AdminDashboardPageSkeleton";
 import { Suspense } from "react";
 
 export default function Page() {
