+++ src/app/(admin)/admin/dashboard/reports/page.tsx
@@ -1,5 +1,5 @@
 import ReportSection from "@/components/dashboard/admin/contents/page/ReportSection";
-import AdminDashboardPageSkeleton from "@/components/skeleton/admin/AdminDashboardPageSkeleton";
+import AdminDashboardPageSkeleton from "@/components/ui/skeleton/admin/AdminDashboardPageSkeleton";
 import { Suspense } from "react";
 
 export default function Page() {
