+++ src/app/(admin)/admin/dashboard/users/page.tsx
@@ -1,5 +1,5 @@
 import UsersSection from "@/components/dashboard/admin/contents/page/UsersSection";
-import AdminDashboardPageSkeleton from "@/components/skeleton/admin/AdminDashboardPageSkeleton";
+import AdminDashboardPageSkeleton from "@/components/ui/skeleton/admin/AdminDashboardPageSkeleton";
 import { Suspense } from "react";
 
 export default function Page() {
