+++ src/app/(user)/dashboard/(mailbox)/receive/page.tsx
@@ -1,6 +1,6 @@
 import LetterDetailModal from "@/components/capsule/detail/LetterDetailModal";
 import MailboxPage from "@/components/dashboard/contents/mailbox/MailboxPage";
-import MailboxSkeleton from "@/components/skeleton/MailboxSkeleton";
+import MailboxSkeleton from "@/components/ui/skeleton/MailboxSkeleton";
 import { Suspense } from "react";
 
 export default async function ReceivePage({
