+++ src/app/(user)/dashboard/storyTrack/joined/page.tsx
@@ -1,5 +1,5 @@
 import JoinedTrackPage from "@/components/dashboard/storyTrack/joined/JoinedTrackPage";
-import StorytrackPageSkeleton from "@/components/skeleton/dashboard/storytrack/StorytrackPageSkeleton";
+import StorytrackPageSkeleton from "@/components/ui/skeleton/dashboard/storytrack/StorytrackPageSkeleton";
 import { Suspense } from "react";
 
 export default function JoinedPage() {
