+++ src/app/(user)/dashboard/storyTrack/mine/page.tsx
@@ -1,5 +1,5 @@
 import MineTrackPage from "@/components/dashboard/storyTrack/mine/MineTrackPage";
-import StorytrackPageSkeleton from "@/components/skeleton/dashboard/storytrack/StorytrackPageSkeleton";
+import StorytrackPageSkeleton from "@/components/ui/skeleton/dashboard/storytrack/StorytrackPageSkeleton";
 import { Suspense } from "react";
 
 export default function MinePage() {
