+++ src/components/dashboard/admin/contents/page/CapsuleSection.tsx
@@ -5,7 +5,7 @@ import AdminHeader from "../AdminHeader";
 import AdminBody from "../body/AdminBody";
 import StatsOverview from "../StatsOverview";
 import { useQuery } from "@tanstack/react-query";
-import AdminDashboardPageSkeleton from "@/components/skeleton/admin/AdminDashboardPageSkeleton";
+import AdminDashboardPageSkeleton from "@/components/ui/skeleton/admin/AdminDashboardPageSkeleton";
 import ApiError from "@/components/common/error/ApiError";
 
 const CAPSULE_TABS = [
