+++ src/components/dashboard/contents/main/second/PendingLetters.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import ApiError from "@/components/common/error/ApiError";
-import PendingLettersSkeleton from "@/components/skeleton/dashboard/home/PendingLettersSkeleton";
+import PendingLettersSkeleton from "@/components/ui/skeleton/dashboard/home/PendingLettersSkeleton";
 import { capsuleDashboardApi } from "@/lib/api/capsule/dashboardCapsule";
 import { formatDateTime } from "@/lib/hooks/formatDateTime";
 import { useQuery } from "@tanstack/react-query";
