+++ src/components/dashboard/contents/main/third/LetterReadingStatus.tsx
@@ -12,7 +12,7 @@ import {
 import DivBox from "../../../DivBox";
 import { useQuery } from "@tanstack/react-query";
 import { capsuleDashboardApi } from "@/lib/api/capsule/dashboardCapsule";
-import LetterReadingSkeleton from "@/components/skeleton/dashboard/home/LetterReadingSkeleton";
+import LetterReadingSkeleton from "@/components/ui/skeleton/dashboard/home/LetterReadingSkeleton";
 import ApiError from "@/components/common/error/ApiError";
 
 export default function LetterReadingStatus() {
