+++ src/components/dashboard/storyTrack/all/AllTrackPage.tsx
@@ -6,7 +6,7 @@ import TrackCard from "./TrackCard";
 import BackButton from "@/components/common/tag/BackButton";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { storyTrackApi } from "@/lib/api/dashboard/storyTrack";
-import StorytrackCardSkeleton from "@/components/skeleton/dashboard/storytrack/StorytrackCardSkeleton";
+import StorytrackCardSkeleton from "@/components/ui/skeleton/dashboard/storytrack/StorytrackCardSkeleton";
 
 type StatusFilter = "all" | "PARTICIPANT" | "NOT_JOINED" | "COMPLETED";
 type SortOption = "newest" | "popular";
