+++ src/components/dashboard/storyTrack/mine/MineTrackPage.tsx
@@ -6,7 +6,7 @@ import StoryMenuTab from "../common/StoryMenuTab";
 import MineCard from "./MineCard";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { storyTrackApi } from "@/lib/api/dashboard/storyTrack";
-import StorytrackCardSkeleton from "@/components/skeleton/dashboard/storytrack/StorytrackCardSkeleton";
+import StorytrackCardSkeleton from "@/components/ui/skeleton/dashboard/storytrack/StorytrackCardSkeleton";
 
 /* 무한스크롤 */
 function useInfiniteScroll(
