+++ src/components/landingPage/section/main/Background.tsx
@@ -3,9 +3,10 @@ export default function Background() {
     <>
       <div className="absolute inset-0 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.13)_99.51%)]"></div>
       <div className="absolute inset-0 bg-white mask-[linear-gradient(180deg,#ffffff_0%,rgba(255,255,255,0)_15%,rgba(255,255,255,0)_85%,#ffffff_100%)]"></div>
-      <div className="absolute top-35 -left-70 rounded-full w-161.75 h-161.75 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.13)_99.51%)] blur-[128px]"></div>
-      <div className="absolute -top-30 right-80 rounded-full w-75 h-75 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.13)_99.51%)] blur-[128px]"></div>
-      <div className="absolute -bottom-12 -right-27 rounded-full w-75 h-75 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.13)_99.51%)] blur-[128px]"></div>
+
+      <div className="absolute top-35 -left-70 rounded-full w-161.75 h-161.75 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.50)_99.51%)] blur-[128px]"></div>
+      <div className="absolute -top-30 right-80 rounded-full w-75 h-75 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.80)_99.51%)] blur-[128px]"></div>
+      <div className="absolute -bottom-12 -right-27 rounded-full w-75 h-75 bg-[linear-gradient(105.07deg,rgba(255,255,255,0.51)_0.49%,rgba(255,218,212,0.3)_50%,rgba(255,137,153,0.80)_99.51%)] blur-[128px]"></div>
     </>
   );
 }
