+++ src/components/ui/skeleton/dashboard/map/MapContentsSkeleton.tsx
