+++ src/components/ui/skeleton/dashboard/map/MapPageSkeleton.tsx
