+++ src/components/ui/skeleton/dashboard/storytrack/StorytrackCardSkeleton.tsx
