+++ src/components/ui/skeleton/dashboard/storytrack/TrackDetailSkeleton.tsx
