+++ src/lib/constants/capsulePalette.ts
